add string vid test case
